+++ CHANGELOG.md
@@ -10,3 +10,4 @@ Please update this document each time you close an issue by your commit.
 - **(enh)** send email on register success
 - **(fix)** _form.php created using gii is fixed
 - **(enh)** recreated documentation bootstrap style
+- **(fix)** cms pages parent branch not displaying issue